Remove unused sessionId, rename authprovider [PAR-25]
